// types/api.types.ts
export type ProductT = {
  id: string;
  name: string;
  description: string | null;
  tags: string | null;
  categories: string;
  brandId: string;
  brandName: string | null;
  likesCount: number;
  isActive: boolean;
  masterImage: string;
  thumbnail: string;
  discountPercent: number;
  masterPrice: number;
  discountPrice: number;
  inStock: number;
  city: string | null;
  code: string | null;
  country: string | null;
  state: string | null;
  type: number;
  commentsCount: number;
  rate: number;
  r1: string | null;
  r2: string | null;
  r3: string | null;
  r4: string | null;
  r5: string | null;
  p1: string | null;
  masterId: string | null;
  createdBy: string;
  jsonExt: string | null;
  attachmentCount: number | null;
  attachmentDuration: number | null;
};